requirements.txt
# Core ML libraries
torch>=1.10.0
tensorflow>=2.8.0
gymnasium>=0.26.3
stable-baselines3>=2.0.0
optuna>=2.10.0
gym>=0.21.0

# Computer Vision (for autonomous interface)
opencv-python>=4.5.0
pytesseract>=0.3.0
pillow>=8.0.0
# tesseract-ocr>=4.0.0  # System package - install separately with apt/brew/chocolatey

# Data processing and visualization
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0
ray>=1.0.0
pytest>=6.0.0
jupyter>=1.0.0
pywin32>=305

# Game interaction
pyautogui>=0.9.53
mss>=6.0.0  # For screen capture
keyboard>=0.13.0  # For keyboard control
pygame>=2.1.0
requests>=2.25.0
pydirectinput>=1.0.4
pynput>=1.7.0 